# simpler 'cpu_scatter_gather_base_kernel'