@nikitaved 2020-03-13
Interesting, while `scatter` shows (sometimes even significant) performance gain, `gather` has some reductions (5-10%). I will try to investigate whether the cause is the loop-callback vs callback-loop order.