@nikitaved 2020-03-18
@ngimel , I did some additional loop optimizations and I observed some significant performance gains in the following "dense indexing" scenario:
```python
import torch
import numpy
from IPython import get_ipython

numpy.random.seed(13)
torch.manual_seed(13)

ipython = get_ipython()

torch.set_num_threads(1)

Ms=256
Ns=512
dim = 0
top_power = 5

for pM in range(top_power):
    M = Ms * (2 ** pM)
    for pN in range(top_power):
        N = Ns * (2 ** pN)
        input_one = torch.rand(M, N)
        index = torch.tensor(numpy.random.randint(0, M, (M, N)))
        print(f"Problem size (MxN): {M}x{N}")
        ipython.magic("timeit torch.gather(input_one, dim, index)")
```
### Before
```
Problem size (MxN): 256x512
762 µs ± 5.17 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
Problem size (MxN): 256x1024
2.12 ms ± 33.4 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 256x2048
3.89 ms ± 9.65 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 256x4096
7.79 ms ± 10.9 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 256x8192
21.2 ms ± 2.96 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
Problem size (MxN): 512x512
979 µs ± 1.33 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
Problem size (MxN): 512x1024
2.98 ms ± 3.12 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 512x2048
8.18 ms ± 58 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 512x4096
21.3 ms ± 214 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
Problem size (MxN): 512x8192
44.5 ms ± 1.04 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
Problem size (MxN): 1024x512
3.52 ms ± 7.07 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 1024x1024
7.74 ms ± 33.8 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 1024x2048
15.2 ms ± 481 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 1024x4096
49.8 ms ± 344 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
Problem size (MxN): 1024x8192
122 ms ± 334 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
Problem size (MxN): 2048x512
7.04 ms ± 7.03 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 2048x1024
19 ms ± 53.6 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 2048x2048
50.8 ms ± 382 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
Problem size (MxN): 2048x4096
172 ms ± 45 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
Problem size (MxN): 2048x8192
363 ms ± 472 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
Problem size (MxN): 4096x512
20.3 ms ± 15.9 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
Problem size (MxN): 4096x1024
71.6 ms ± 282 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
Problem size (MxN): 4096x2048
179 ms ± 1.19 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
Problem size (MxN): 4096x4096
379 ms ± 622 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
Problem size (MxN): 4096x8192
803 ms ± 306 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

### After
```
Problem size (MxN): 256x512                                                                                                                       
145 µs ± 1.1 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)                                                                           
Problem size (MxN): 256x1024                                                                                                                      
441 µs ± 410 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)                                                                            
Problem size (MxN): 256x2048                                                                                                                      
963 µs ± 1.09 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)                                                                           
Problem size (MxN): 256x4096                                                                                                                      
2.87 ms ± 6.12 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)                                                                           
Problem size (MxN): 256x8192                                                                                                                      
11.9 ms ± 5.67 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)                                                                           
Problem size (MxN): 512x512                                                                                                                       
328 µs ± 709 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)                                                                            
Problem size (MxN): 512x1024                                                                                                                      
998 µs ± 563 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)                                                                            
Problem size (MxN): 512x2048                                                                                                                      
2.53 ms ± 3.41 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 512x4096
11.7 ms ± 136 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 512x8192
33.5 ms ± 31.6 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
Problem size (MxN): 1024x512
4.31 ms ± 6.15 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 1024x1024
3.03 ms ± 3.21 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 1024x2048
11.2 ms ± 19 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 1024x4096
31 ms ± 130 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
Problem size (MxN): 1024x8192
85 ms ± 470 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
Problem size (MxN): 2048x512
7.74 ms ± 30.4 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 2048x1024
20.4 ms ± 27.9 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
Problem size (MxN): 2048x2048
28.6 ms ± 25.2 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
Problem size (MxN): 2048x4096
79.6 ms ± 327 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
Problem size (MxN): 2048x8192
166 ms ± 346 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
Problem size (MxN): 4096x512
22.4 ms ± 22.1 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
Problem size (MxN): 4096x1024
72.1 ms ± 248 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
Problem size (MxN): 4096x2048
179 ms ± 103 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
Problem size (MxN): 4096x4096
158 ms ± 35 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
Problem size (MxN): 4096x8192
324 ms ± 487 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)

```

The rest is about 2-5% faster or very similar.