@ngimel 2020-03-18
This is impressive! Do the gains hold for both scatter and gather? I'll review and merge soon, thank you!