@nikitaved 2020-03-18
scatter is less impressive, maybe by a factor of about 2x in the "dense indexing" test scenario. I am not sure that the heuristic that I use for this very simple loop optimization is optimal, however, need some expert's opinion.